Make stuff more generic and move to separate module.
